[긴급][기능개선][엔티티] 게시글 엔티티 추상 클래스 (BaseEntity, BasePost) 리팩토링: refactor: baseEntity, basePost 추상 클래스 수정에 따른 게시글 클래스 엔티티 리팩토링 #746
